[refactor] class定義をそれぞれ1ファイルにする
