+++ src/models/PodcastStationManager.ts
@@ -1,4 +1,5 @@
 import BaseModel from './BaseModel';
+import PodcastStation from './PodcastStation';
 
 export default class PodcastStationManager extends BaseModel {
   constructor() {
@@ -42,26 +43,5 @@ export default class PodcastStationManager extends BaseModel {
     console.log(list);
     this.saveStore(list);
     console.log(this.geAll());
-  }
-}
-
-export class PodcastStation {
-  rssUrl: string;
-  name: string;
-
-  constructor(name: string, rssUrl: string) {
-    this.name = name;
-    this.rssUrl = rssUrl;
-  }
-
-  /**
-   *to_json
-   */
-  public toDictionary() {
-    return { name: this.name, rssUrl: this.rssUrl };
-  }
-
-  static fromDictionary(dic: { name: string; rssUrl: string }) {
-    return new PodcastStation(dic.name, dic.rssUrl);
   }
 }
